[Servers] Credentials storage: Add a simple file keyring with no encryption
Used as fallback when no Keyring backends are found (Secret service, 
KWallet, ...)

This change will also allow the addition of other custom keyring 
backends: for ex. support of external password managers

TODO: Ask user if passwords should be saved in plaintext or if they 
shouldn't be saved at all

Later if external password managers are available, add an option in a 
new Preferences page called "Advanced" to allow the user to choose one.