[Reader] Partially fix issues #131 and #141
Unfortunately, the Cairo library suffers from a limitation on the size 
of the images (32k pixels max.). Its use is necessary for rendering 
images with screens with high DPI.

With this fix, when the DPI scale is 1, Gtk.Image are filled using 
directly the Gdk.Pixbuf rather than a Cairo.Surface.

The problem remains for screens with high DPI when images are too big, 
which is frequent with Webtoons.